Add files via upload
These files are a major revision to how I handled passing the data from the APIs.  The original version just took in the JSON raw and passed it off to the user of the .py classes.  In this version, I created a set of classes that convert the API information into proper objects.